Move Spark2 implementation of StreamingOffset from iceberg-spark to iceberg-Spark2
